// my-tiktuk/src/types/video.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface GetUserFeed {
    statusCode: number;
    itemList:   ItemList[];
    cursor:     string;
    hasMore:    boolean;
}

export interface ItemList {
    id:                string;
    desc:              string;
    createTime:        number;
    video:             Video;
    author:            Author;
    music:             Music;
    challenges:        Challenge[];
    stats:             Stats;
    duetInfo:          DuetInfo;
    originalItem:      boolean;
    officalItem:       boolean;
    textExtra:         TextExtra[];
    secret:            boolean;
    forFriend:         boolean;
    digged:            boolean;
    itemCommentStatus: number;
    showNotPass:       boolean;
    vl1:               boolean;
    itemMute:          boolean;
    effectStickers?:   EffectSticker[];
    authorStats:       AuthorStats;
    privateItem:       boolean;
    duetEnabled:       boolean;
    stitchEnabled:     boolean;
    shareEnabled:      boolean;
    isAd:              boolean;
    duetDisplay:       number;
    stitchDisplay:     number;
    stickersOnItem?:   StickersOnItem[];
    warnInfo?:         WarnInfo[];
}

export interface Author {
    id:             string;
    uniqueId:       string;
    nickname:       string;
    avatarThumb:    string;
    avatarMedium:   string;
    avatarLarger:   string;
    signature:      string;
    verified:       boolean;
    secUid:         string;
    secret:         boolean;
    ftc:            boolean;
    relation:       number;
    openFavorite:   boolean;
    commentSetting: number;
    duetSetting:    number;
    stitchSetting:  number;
    privateAccount: boolean;
}

export interface AuthorStats {
    followingCount: number;
    followerCount:  number;
    heartCount:     number;
    videoCount:     number;
    diggCount:      number;
    heart:          number;
}

export interface Challenge {
    id:            string;
    title:         string;
    desc:          string;
    profileThumb:  string;
    profileMedium: string;
    profileLarger: string;
    coverThumb:    string;
    coverMedium:   string;
    coverLarger:   string;
    isCommerce:    boolean;
}

export interface DuetInfo {
    duetFromId: string;
}

export interface EffectSticker {
    name: string;
    ID:   string;
}

export interface Music {
    id:          string;
    title:       string;
    playUrl:     string;
    coverThumb:  string;
    coverMedium: string;
    coverLarge:  string;
    authorName:  string;
    original:    boolean;
    duration:    number;
    album:       string;
}

export interface Stats {
    diggCount:    number;
    shareCount:   number;
    commentCount: number;
    playCount:    number;
}

export interface StickersOnItem {
    stickerType: number;
    stickerText: string[];
}

export interface TextExtra {
    awemeId:      string;
    start:        number;
    end:          number;
    hashtagName:  string;
    hashtagId:    string;
    type:         number;
    userId:       string;
    isCommerce:   boolean;
    userUniqueId: string;
    secUid:       string;
    subType:      number;
}

export interface Video {
    id:            string;
    height:        number;
    width:         number;
    duration:      number;
    ratio:         string;
    cover:         string;
    originCover:   string;
    dynamicCover:  string;
    playAddr:      string;
    downloadAddr:  string;
    shareCover:    string[];
    reflowCover:   string;
    bitrate:       number;
    encodedType:   string;
    format:        string;
    videoQuality:  string;
    encodeUserTag: string;
    codecType:     string;
    definition:    string;
}

export interface WarnInfo {
    text: string;
    url:  string;
    type: number;
    lang: string;
    key:  string;
}